+++ molo/core/tasks.py
@@ -178,7 +178,7 @@ def rotate_featured_in_homepage(main_lang, day):
                             random_article = ArticlePage.objects.live().filter(
                                 featured_in_homepage=False,
                                 languages__language__id=main_lang.id
-                            ).child_of(section).order_by('?').exact_type(
+                            ).descendant_of(section).order_by('?').exact_type(
                                 ArticlePage).first()
 
                             # promotes an article and bumps last one off list
@@ -235,3 +235,16 @@ def validate_content(data, locales, username, email, host):
         'errors': result['errors'],
         'warnings': result['warnings']
     })
+
+
+@task(ignore_result=True)
+def molo_consolidated_minute_task():
+    """
+    Purpose: To reduce the number of db connections, we run all
+    tasks in 1 call
+    """
+
+    demote_articles()
+    promote_articles()
+    publish_scheduled_pages()
+    clearsessions()